Bugfix: allow separate debounce intervals for push and release.
The "hold for N seconds" pattern used for 'exit admin mode' and 'reboot'
was actually "hold for N seconds and then wait for it to be released
for N seconds" which is a weird and unintuitive operation, since you
could interleave more buttons in the release idle period.

Now a lengthy debounce for push and a quick debounce for release
make the buttons do what I want.