Bugfix: EF_BUILD_RANDOM buffer overrun disables I2C buttons.
The Animation parameter initialization for `Effect::EF_BUILD_RANDOM`
cleared out a memory array using `memset()` that zeroed out far past
the end of the array.

(At least one of) the effects of this is that the I2C communication
with the 8 I2C-based buttons on the button grid would silently fail
for the remainder of the sketch execution, after this animation started.

Reset `memset()` size parameter to use correct type (uint8_t) for the
`Animation::_buildRandomOrder` buffer.